메인페이지에서 최근 등록된 상품 조회

+++ src/main/java/com/project/tuni_back/bean/vo/UserVO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class UserVO {
 	private String userId;
-	private int schoolId;
+	private Long schoolId;
 	private String email;
 }
